event status fix
