+++ pages/events/admin.js
@@ -2,40 +2,10 @@ import {useState, useEffect} from "react";
 import { useSelector, useDispatch } from 'react-redux'
 import { useRouter } from 'next/router'
 import Link from "next/link";
+import EventCard from "../../components/Event/Card";
 
 // blockquote ないかも
 import {Card, Row, Col, Container, blockquote} from "react-bootstrap";
-
-const EventCard = (props) => {
-
-  const { event } = props
-
-  return <Row className={"mt-3"}>
-    <Col>
-      <Link href={`/events/${event.id}`}>
-        <Card>
-          <Card.Header>{event.title}</Card.Header>
-          <Card.Body>
-            <blockquote className="blockquote mb-0">
-              <p>
-                {event.description}
-              </p>
-              <footer className={'blockquote-footer'} >
-                { event.date }
-              </footer>
-              <footer className="blockquote-footer">
-                { event.startTime }~{ event.endTime }
-              </footer>
-              <footer className="blockquote-footer">
-                イベント作成者: { event.administrator.name }
-              </footer>
-            </blockquote>
-          </Card.Body>
-        </Card>
-      </Link>
-    </Col>
-  </Row>
-}
 
 const Index = (props) => {
 
@@ -95,7 +65,9 @@ const Index = (props) => {
 
   return <>
     {
-      events.map(event => <EventCard event={event} key={event.id} />)
+      events.map(event => <Link key={event.id} href={`/events/${event.id}`}>
+        <EventCard event={event} />
+      </Link>)
     }
   </>
 }
