+++ pages/users/events.js
@@ -3,39 +3,7 @@ import { useSelector, useDispatch } from 'react-redux'
 import { useRouter } from 'next/router'
 import Link from "next/link";
 
-// blockquote ないかも
-import {Card, Row, Col, Container, blockquote} from "react-bootstrap";
-
-const EventCard = (props) => {
-
-  const { event } = props
-
-  return <Row className={"mt-3"}>
-    <Col>
-      <Link href={`/events/${event.id}`}>
-        <Card>
-          <Card.Header>{event.title}</Card.Header>
-          <Card.Body>
-            <blockquote className="blockquote mb-0">
-              <p>
-                {event.description}
-              </p>
-              <footer className={'blockquote-footer'} >
-                { event.date }
-              </footer>
-              <footer className="blockquote-footer">
-                { event.startTime }~{ event.endTime }
-              </footer>
-              <footer className="blockquote-footer">
-                イベント作成者: { event.administrator.name }
-              </footer>
-            </blockquote>
-          </Card.Body>
-        </Card>
-      </Link>
-    </Col>
-  </Row>
-}
+import EventCard from "../../components/Event/Card";
 
 const Index = (props) => {
 
@@ -87,7 +55,9 @@ const Index = (props) => {
 
   return <>
     {
-      events.map(event => <EventCard event={event} key={event.id} />)
+      events.map(event => <Link key={event.id} href={`/events/${event.id}`} >
+        <EventCard event={event} key={event.id} />
+        </>)
     }
   </>
 }
